Add achievements plugin support for organization account (#230)
