+++ source/app/web/statics/about/index.html
@@ -37,7 +37,7 @@
               </button>
             </div>
             <small class="info">
-              Display rankings, highlights, commits calendar, used languages and recent activity from any user account!
+              Display rankings, contributions, highlights, commits calendar, used languages and recent activity from any user account!
             </small>
             <small class="info" v-if="metrics">
               Share this profile using <a :href="url">{{ url }}</a>
@@ -66,7 +66,7 @@
             </div>
           </section>
 
-          <div class="rankeds" v-if="account.type === 'user'">
+          <div class="rankeds">
             <div v-for="{icon, title, text, rank, progress, value, leaderboard = null} in ranked" class="ranked" :class="{[rank.charAt(0).toLocaleLowerCase()]:rank !== '$', secret:rank === '$'}">
               <div class="icon">
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 60 60" height="44" width="44">
@@ -104,7 +104,7 @@
             </div>
           </section>
 
-          <section class="container" v-if="account.type === 'user'">
+          <section class="container">
             <h2>
               <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" width="16" height="16"><path d="M8.5.75a.75.75 0 00-1.5 0v5.19L4.391 3.33a.75.75 0 10-1.06 1.061L5.939 7H.75a.75.75 0 000 1.5h5.19l-2.61 2.609a.75.75 0 101.061 1.06L7 9.561v5.189a.75.75 0 001.5 0V9.56l2.609 2.61a.75.75 0 101.06-1.061L9.561 8.5h5.189a.75.75 0 000-1.5H9.56l2.61-2.609a.75.75 0 00-1.061-1.06L8.5 5.939V.75z"></path></svg>
               Highlights
