+++ source/plugins/achievements/metadata.yml
@@ -4,6 +4,7 @@ categorie: github
 index: 17
 supports:
   - user
+  - organization
 inputs:
 
   # Enable or disable plugin
